Fix narrowing conversion warning
Fix the following warning by casting to the appropriate timeval member types:

In file included from /home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/Encryption.hpp:5,
                 from /home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/WBTransmitter.h:25,
                 from /home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/WBTransmitter.cpp:19:
/home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/HelperSources/Helper.hpp: In function ‘constexpr timeval GenericHelper::durationToTimeval(std::chrono::nanoseconds)’:
/home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/HelperSources/Helper.hpp:212:40: warning: narrowing conversion of ‘us.std::chrono::duration<long long int, std::ratio<1, 1000000> >::count()’ from ‘std::chrono::duration<long long int, std::ratio<1, 1000000> >::rep’ {aka ‘long long int’} to ‘__suseconds_t’ {aka ‘long int’} [-Wnarrowing]
  212 |   return timeval{secs.count(), us.count()};
      |                                ~~~~~~~~^~

This warning is not addressed, as 32-bit time_t will overflow in 2038, and will need fixed:
  /home/openhd/code/OpenHD/OpenHD/lib/wifibroadcast/src/HelperSources/Helper.hpp:212:28: warning: narrowing conversion of ‘secs.std::chrono::duration<long long int>::count()’ from ‘std::chrono::duration<long long int>::rep’ {aka ‘long long int’} to ‘__time_t’ {aka ‘long int’} [-Wnarrowing]
    212 |   return timeval{secs.count(), us.count()};
        |                  ~~~~~~~~~~^~

+++ src/HelperSources/Helper.hpp
@@ -209,7 +209,7 @@ static constexpr timeval durationToTimeval(nanoseconds dur) {
   const auto secs = duration_cast<seconds>(dur);
   dur -= secs;
   const auto us = duration_cast<microseconds>(dur);
-  return timeval{secs.count(), us.count()};
+  return timeval{secs.count(), (long int)us.count()};
 }
 }
